+++ packages/components/ContextualMenu/CHANGELOG.md
@@ -1,8 +1,38 @@
 # Change Log - @fluentui-react-native/contextual-menu
 
-This log was last generated on Thu, 06 Jan 2022 23:14:38 GMT and should not be manually modified.
+This log was last generated on Fri, 14 Jan 2022 21:49:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.14.11
+
+Fri, 14 Jan 2022 21:49:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.19.16
+- Bump @fluentui-react-native/focus-zone to v0.8.13
+- Bump @fluentui-react-native/icon to v0.11.5
+- Bump @fluentui-react-native/interactive-hooks to v0.11.13
+- Bump @fluentui-react-native/text to v0.11.8
+- Bump @uifabricshared/foundation-compose to v1.11.8
+- Bump @fluentui-react-native/pressable to v0.8.13
+
+## 0.14.10
+
+Fri, 14 Jan 2022 01:00:01 GMT
+
+### Patches
+
+- fix disable buttonn (email not defined)
+- Bump @fluentui-react-native/callout to v0.19.15
+- Bump @fluentui-react-native/focus-zone to v0.8.12
+- Bump @fluentui-react-native/icon to v0.11.4
+- Bump @fluentui-react-native/interactive-hooks to v0.11.12
+- Bump @fluentui-react-native/text to v0.11.7
+- Bump @fluentui-react-native/tokens to v0.11.5
+- Bump @uifabricshared/foundation-compose to v1.11.7
+- Bump @fluentui-react-native/pressable to v0.8.12
 
 ## 0.14.9