+++ packages/deprecated/theming-react-native/CHANGELOG.md
@@ -1,8 +1,28 @@
 # Change Log - @uifabricshared/theming-react-native
 
-This log was last generated on Thu, 06 Jan 2022 23:14:38 GMT and should not be manually modified.
+This log was last generated on Fri, 14 Jan 2022 21:49:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.9
+
+Fri, 14 Jan 2022 21:49:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.9.8
+- Bump @fluentui-react-native/win32-theme to v0.14.9
+
+## 0.13.8
+
+Fri, 14 Jan 2022 01:00:04 GMT
+
+### Patches
+
+- fix disabled button (email not defined)
+- Bump @fluentui-react-native/default-theme to v0.9.7
+- Bump @fluentui-react-native/win32-theme to v0.14.8
+- Bump @uifabricshared/theming-ramp to v0.16.4
 
 ## 0.13.7