+++ packages/experimental/Button/CHANGELOG.md
@@ -1,8 +1,46 @@
 # Change Log - @fluentui-react-native/experimental-button
 
-This log was last generated on Fri, 07 Jan 2022 21:27:06 GMT and should not be manually modified.
+This log was last generated on Tue, 18 Jan 2022 20:49:22 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.17
+
+Tue, 18 Jan 2022 20:49:22 GMT
+
+### Patches
+
+- Font fixes (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/styling-utils to v0.3.4
+
+## 0.11.16
+
+Fri, 14 Jan 2022 21:49:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-activity-indicator to v0.5.10
+- Bump @fluentui-react-native/experimental-text to v0.7.11
+- Bump @fluentui-react-native/framework to v0.7.11
+- Bump @fluentui-react-native/icon to v0.11.5
+- Bump @fluentui-react-native/interactive-hooks to v0.11.13
+- Bump @fluentui-react-native/theme-tokens to v0.11.6
+
+## 0.11.15
+
+Fri, 14 Jan 2022 01:00:02 GMT
+
+### Patches
+
+- fix disabled button (email not defined)
+- Bump @fluentui-react-native/experimental-activity-indicator to v0.5.9
+- Bump @fluentui-react-native/experimental-text to v0.7.10
+- Bump @fluentui-react-native/framework to v0.7.10
+- Bump @fluentui-react-native/icon to v0.11.4
+- Bump @fluentui-react-native/interactive-hooks to v0.11.12
+- Bump @fluentui-react-native/styling-utils to v0.3.3
+- Bump @fluentui-react-native/theme-tokens to v0.11.5
+- Bump @fluentui-react-native/tokens to v0.11.5
 
 ## 0.11.14