+++ packages/experimental/Checkbox/CHANGELOG.md
@@ -1,8 +1,42 @@
 # Change Log - @fluentui-react-native/experimental-checkbox
 
-This log was last generated on Fri, 07 Jan 2022 23:38:55 GMT and should not be manually modified.
+This log was last generated on Wed, 19 Jan 2022 00:57:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.0
+
+Wed, 19 Jan 2022 00:57:04 GMT
+
+### Minor changes
+
+- Initial styling (ruaraki@microsoft.com)
+
+## 0.5.2
+
+Fri, 14 Jan 2022 21:49:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-text to v0.7.11
+- Bump @fluentui-react-native/framework to v0.7.11
+- Bump @fluentui-react-native/interactive-hooks to v0.11.13
+- Bump @fluentui-react-native/pressable to v0.8.13
+- Bump @fluentui-react-native/theme-tokens to v0.11.6
+
+## 0.5.1
+
+Fri, 14 Jan 2022 01:00:02 GMT
+
+### Patches
+
+- fix disabled button (email not defined)
+- Bump @fluentui-react-native/experimental-text to v0.7.10
+- Bump @fluentui-react-native/framework to v0.7.10
+- Bump @fluentui-react-native/interactive-hooks to v0.11.12
+- Bump @fluentui-react-native/pressable to v0.8.12
+- Bump @fluentui-react-native/theme-tokens to v0.11.5
+- Bump @fluentui-react-native/tokens to v0.11.5
 
 ## 0.5.0