+++ packages/theming/theme-tokens/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @fluentui-react-native/theme-tokens
 
-This log was last generated on Thu, 06 Jan 2022 23:14:38 GMT and should not be manually modified.
+This log was last generated on Fri, 14 Jan 2022 21:49:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.6
+
+Fri, 14 Jan 2022 21:49:07 GMT
+
+### Patches
+
+- resolve json modules in tsconfig (sanajmi@microsoft.com)
+
+## 0.11.5
+
+Fri, 14 Jan 2022 01:00:03 GMT
+
+### Patches
+
+- fix disabled button (email not defined)
+- Bump @fluentui-react-native/theme-types to v0.13.3
 
 ## 0.11.4