+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Thu, 06 Jan 2022 23:14:38 GMT and should not be manually modified.
+This log was last generated on Fri, 14 Jan 2022 21:49:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.13
+
+Fri, 14 Jan 2022 21:49:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.11.5
+
+## 0.11.12
+
+Fri, 14 Jan 2022 01:00:03 GMT
+
+### Patches
+
+- fix disabled button (email not defined)
+- Bump @fluentui-react-native/icon to v0.11.4
 
 ## 0.11.11